pqc_iov_auth/requirements.txt
cryptography>=41.0.0
liboqs-python>=0.8.0
redis>=4.5.0
requests>=2.31.0
pytest>=7.0.0
pytest-mock>=3.10.0